socket-activate printer
Previous to this commit, the printer needs to continuously run in the
background, even if it doesn't do much most of the time, to keep the
socket open and handle incoming connections. This commit hands over
handling the open socket to systemd and the printer-code gets started by
systemd only when there is something incoming on the socket. The printer
itself then takes over the socket from systemd, handles the incoming
data and exits afterwards, returning socket management to systemd again.

This reduces running processes as well as the memory footprint of the
running system.